CheerRepository 위치이동
