+++ app/src/main/java/dev/kichan/marketplace/model/repository/CheerRepository.kt
@@ -1,0 +1,12 @@
+package dev.kichan.marketplace.model.repository
+
+import dev.kichan.marketplace.model.service.CheerService
+import dev.kichan.marketplace.model.NetworkModule
+
+class CheerRepository {
+    val service = NetworkModule.getService(CheerService::class.java)
+
+    suspend fun cheer(
+        marketId: Long
+    ) = service.cheer(marketId)
+}
